+++ apps/mobile-poc/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @moxxy/mobile-poc
 
+## 0.0.14
+
+### Patch Changes
+
+- @moxxy/client-core@0.8.1
+- @moxxy/client-transport-ws@0.1.17
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [c058735]
+  - @moxxy/client-core@0.8.0
+  - @moxxy/client-transport-ws@0.1.16
+
 ## 0.0.12
 
 ### Patch Changes