+++ packages/client-platform-web/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @moxxy/client-platform-web
 
+## 0.1.18
+
+### Patch Changes
+
+- @moxxy/client-core@0.8.1
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [c058735]
+  - @moxxy/client-core@0.8.0
+
 ## 0.1.16
 
 ### Patch Changes