+++ packages/ipc-server-ws/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @moxxy/ipc-server-ws
 
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [f8b0c63]
+  - @moxxy/desktop-ipc-contract@0.9.0
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [c058735]
+  - @moxxy/desktop-ipc-contract@0.8.0
+
 ## 0.1.15
 
 ### Patch Changes